--- asapdiscovery/postera_api/requirements.txt
@@ -1,3 +1,0 @@
-typing_extensions==4.0.0
-pandas==1.2.3
-requests==2.21.0
